modify Task_02
