+++ 2020-Autumn/Backend/task_02/huoweihao/t3.cpp
@@ -1,0 +1,13 @@
+#include<iostream>
+using namespace std;
+//3.±éÀúÊı×é
+void traversal1(int* a, int n)
+{
+    for (int i = 0; i < n; i++)
+        cout << *(a + i) << "	";
+}
+int main()
+{
+
+	return 0;
+}
